Moved test classes
	modified:   base/tests/views/education_groups/test_detail.py
	modified:   base/tests/views/test_education_group.py
